Add: Build script for  dbg plus small changes
git-svn-id: https://neolua.svn.codeplex.com/svn/trunk@31920 2abf6ff4-378a-4344-b6f7-2bcd8b316ba9

+++ NeoLua.Test/Expressions.cs
@@ -697,7 +697,7 @@ namespace LuaDLR.Test
     [TestMethod]
     public void TestArithmetic23()
     {
-      TestCode("return -0x80", (short)-0x80);
+      TestCode("return -0x80", (int)-0x80);
       TestCode("return -0x8000", (int)-0x8000);
       TestCode("return -0x80000000", (long)-0x80000000);
     }
